Refactor: delete folder for code 1.step2 폴더 삭제 -아이스크림 출력 코드와 빼빼로 출력코드를 한 파일로 합쳐 불필요한 부분을 수정했습니다.
-
